+++ RenderObject.java
@@ -7,7 +7,6 @@
  */
 
 import javafx.scene.image.Image;
-import java.math.BigDecimal;
 
 /**
  * RenderObject defines objects which can be rendered, they all sprites and position in common.
@@ -37,12 +36,12 @@ public abstract class RenderObject {
 	 * getter for position
 	 * splits into x and y array instead of just a position object.
 	 */
-	public BigDecimal[] getObjectPosition() {
+	public double[] getObjectPosition() {
 		return this.objectPosition.getPosition();
 	}
 	
 	//setter for object position 
-	public void setObjectPosition(BigDecimal x, BigDecimal y) {
+	public void setObjectPosition(double x, double y) {
 		this.objectPosition.setPosition(x, y);
 	}
